% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folk.R
\name{ethno_music_stat}
\alias{ethno_music_stat}
\title{Focusing on providing high quality music streaming content}
\usage{
ethno_music_stat(type, ...)
}
\arguments{
\item{type}{is the ethnical and cultural inspired music of all genres and all regions of the world. Example \code{\link{ethno_music_type}()}}
}
\description{
Focusing on providing high quality music streaming content
}
\author{
Jean Marie Cimula
}

