% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folk.R
\name{ethno_music_stat_all}
\alias{ethno_music_stat_all}
\title{Focusing on providing high quality music streaming content}
\usage{
ethno_music_stat_all()
}
\description{
Focusing on providing high quality music streaming content
}

