% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variety.R
\name{ethno_music_type}
\alias{ethno_music_type}
\title{Focusing on providing high quality music streaming content}
\usage{
ethno_music_type()
}
\description{
Focusing on providing high quality music streaming content
}
\details{
List of all regions of the world
}

